% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{MyHiker}
\alias{MyHiker}
\alias{$.MyHiker}
\title{show case persist memory between calls to rust}
\format{
An object of class \code{environment} of length 2.
}
\usage{
MyHiker
}
\description{
show case persist memory between calls to rust
}
\keyword{datasets}
